fix(cli): preserve MCP symlinks and JSON numeric source text
